instrumentation/opentelemetry-instrumentation-tortoiseorm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-tortoiseorm"
dynamic = ["version"]
description = "OpenTelemetry Instrumentation for Tortoise ORM"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.7"
authors = [
  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "opentelemetry-api ~= 1.12",
  "opentelemetry-instrumentation == 0.36b0.dev",
  "opentelemetry-semantic-conventions == 0.36b0.dev",
]

[project.optional-dependencies]
instruments = [
  "tortoise-orm >= 0.17.0",
  "pydantic >= 1.10.2"
]
test = [
  "opentelemetry-instrumentation-tortoiseorm[instruments]",
  "opentelemetry-test-utils == 0.36b0.dev",
]

[project.entry-points.opentelemetry_instrumentor]
tortoiseorm = "opentelemetry.instrumentation.tortoiseorm:TortoiseORMInstrumentor"

[project.urls]
Homepage = "https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation/opentelemetry-instrumentation-tortoiseorm"

[tool.hatch.version]
path = "src/opentelemetry/instrumentation/tortoiseorm/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]
